Crud completed

+++ patient-management-frontend/src/app/components/patient-list/patient-list.component.ts
@@ -1,15 +1,25 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild, NgModule } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { PatientService } from '../../service/patient.service';
-import { BrowserModule } from '@angular/platform-browser';
 import { MatTableModule } from '@angular/material/table';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatSortModule } from '@angular/material/sort';
 import { MatInputModule } from '@angular/material/input';
 import { MatToolbarModule } from '@angular/material/toolbar';
-
+import {MatIconModule} from '@angular/material/icon';
+import {MatDialog} from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { AddPatientDialogComponent } from '../add-patient-dialog/add-patient-dialog.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
+import { MatDialogModule } from '@angular/material/dialog';
+import * as XLSX from 'xlsx';
+import { EditPatientDialogComponent } from '../edit-patient-dialog/edit-patient-dialog.component';
+import { Patient } from '../../models/patient.model';
 
 @Component({
   selector: 'app-patient-list',
@@ -19,19 +29,28 @@ import { MatToolbarModule } from '@angular/material/toolbar';
     MatPaginatorModule,
     MatSortModule,
     MatInputModule,
-    MatToolbarModule
+    MatToolbarModule,
+    MatIconModule, 
+    ReactiveFormsModule, 
+    MatButtonModule, 
+    MatFormFieldModule,
+    MatSelectModule, 
+    MatDialogModule
   ],
   templateUrl: './patient-list.component.html',
   styleUrls: ['./patient-list.component.css']
 })
 export class PatientListComponent implements OnInit {
-  displayedColumns: string[] = ['N_Dossier', 'CNIE', 'nom', 'prenom', 'Date_Naissance', 'sexe', 'provenance', 'niveauDeScolarite', 'couverture', 'action'];
+  displayedColumns: string[] = ['action', 'N_Dossier', 'CNIE', 'nom', 'prenom', 'Date_Naissance', 'sexe', 'provenance', 'niveauDeScolarite', 'couverture'];
   dataSource!: MatTableDataSource<any>;
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  constructor(private patientService: PatientService) { }
+  constructor(
+    private patientService: PatientService,
+    public dialog: MatDialog
+  ) { }
 
   ngOnInit(): void {
     this.retrievePatients();
@@ -61,13 +80,62 @@ export class PatientListComponent implements OnInit {
   }
 
   deletePatient(N_Dossier: any): void {
-    this.patientService.delete(N_Dossier)
-      .subscribe(
-        response => {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.patientService.delete(N_Dossier)
+          .subscribe(
+            response => {
+              this.retrievePatients(); // Refresh the list after deletion
+            },
+            error => {
+              console.log(error);
+            });
+      }
+    });
+  }
+  openAddPatientDialog(): void {
+    const dialogRef = this.dialog.open(AddPatientDialogComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.patientService.create(result).subscribe(() => {
           this.retrievePatients();
-        },
-        error => {
-          console.log(error);
         });
+      }
+    });
+  }
+
+  openModifyPatientDialog(patient: Patient): void {
+    console.log('Patient data:', patient);
+    const dialogRef = this.dialog.open(EditPatientDialogComponent, {
+      data: patient
+    });
+  
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.updatePatient(result);
+      }
+    });
+  }
+
+  updatePatient(patient:Patient): void{
+    this.patientService.update(patient.N_Dossier, patient).subscribe(() => {
+      this.retrievePatients(); // Refresh the list after update
+    });
+  }
+
+  fileName= "ExcelPatients.xlsx";
+
+  exportexcel(){
+  if (!this.dataSource || !this.dataSource.data) {
+    console.error('No data available for export');
+    return;
+  }
+  const data = this.dataSource.data;
+  const ws: XLSX.WorkSheet = XLSX.utils.json_to_sheet(data);
+  const wb: XLSX.WorkBook = XLSX.utils.book_new();
+  XLSX.utils.book_append_sheet(wb, ws, 'sheet1');
+  XLSX.writeFile(wb, this.fileName);
   }
 }
